+++ Examples/HelloWorld/Structs.cs
@@ -1,0 +1,19 @@
+namespace Samples
+{
+	public struct Position
+	{
+		public float X;
+		public float Y;
+	}
+
+	public struct Velocity
+	{
+		public float X;
+		public float Y;
+	}
+
+	public struct Speed
+	{
+		public int SpeedValue;
+	}
+}
